+++ server/src/services/services.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { BackblazeB2Service } from './backblaze-b2.service';
+import { CloudflareR2Service } from './cloudflare-r2.service';
 
 @Module({
   imports: [ConfigModule],
-  providers: [BackblazeB2Service],
-  exports: [BackblazeB2Service],
+  providers: [CloudflareR2Service],
+  exports: [CloudflareR2Service],
 })
 export class ServicesModule {}